index.html: Sample app that opens a window
It's important to do it this way rather than with an `index.js` because the
latter doesn't open a window (unless you code it explicitly) and that makes
it hard to exit. Just sending an INT signal will bring you back to the
command prompt, but that seems to leave a lot of nw background processes
running that I needed to chase down and kill by hand.

The package.json changes and index.html came from the Getting Started[1]
document, though I start the app with the the `npm start` script as
described the GitLab issue.

[1]: https://nwjs.readthedocs.io/en/latest/For%20Users/Getting%20Started/